🧹 Removed old `MaskAnnotator` implementation as it is no longer used. Moved `ColorMap`, `resolve_color_idx`, `resolve_color` and `BaseAnnotator` to `supervision/annotators/base`. Removed `BoxMaskAnnotator` as it can be achieved by `BoundingBoxAnnotator` and `MaskAnnotator` composition.
